add test_read_token_cache test, and move all tests in oauth2.rs to test_oauth2.rs
